Updated to address issues #1 and #6
Regarding #1, there is now a hidden element added to the form (if your
element has the name “myelem”, then it will create a hidden field with
the name “myelem_pass”, which will hold either “true” or “false” for
server-side password checking.
For issue #6, the minAcceptableScore is now 0-100 to match the pass
value.
Code has also been rearranged to hopefully make it a little more
manageable.